+++ Circles/Views/Photo Galleries/PhotosUploadView.swift
@@ -64,15 +64,51 @@ struct PhotosUploadView: View {
         .padding(10)
         .onAppear {
             task = task ?? UploadTask(priority: .background) {
+                
+                CirclesApp.logger.debug("Uploading \(items.count) items")
 
                 while !items.isEmpty {
                     currentItem = items.removeFirst()
-                    if let data = try? await currentItem?.loadTransferable(type: Data.self),
-                       let img = UIImage(data: data) {
+                    
+                    guard let item = currentItem
+                    else { break }
+                    
+                    if item.supportedContentTypes.contains(where: {
+                        CirclesApp.logger.debug("Checking whether \($0) is a movie type")
+                        return $0.isSubtype(of: .movie)
+                    }) {
+                        // Load it as a movie
+                        CirclesApp.logger.debug("Loading movie")
+                        if let movie = try? await item.loadTransferable(type: Movie.self) {
+                            CirclesApp.logger.debug("Successfully Loaded movie")
+                            if !Task.isCancelled {
+                                let thumbnail = try await movie.loadThumbnail()
+                                await MainActor.run {
+                                    self.currentImage = thumbnail
+                                }
+                                let url = movie.url
+                                try await room.sendVideo(url: url, thumbnail: thumbnail)
+                            }
+                        } else {
+                            // FIXME: Set error message
+                            CirclesApp.logger.error("Failed to load movie")
+                        }
+                    }
+                    else if item.supportedContentTypes.contains(where: {
+                        CirclesApp.logger.debug("Checking whether \($0) is an image type")
+                        return $0.isSubtype(of: .image)
+                    }),
+                            let data = try? await currentItem?.loadTransferable(type: Data.self),
+                            let img = UIImage(data: data)
+                    {
+                        CirclesApp.logger.debug("Loading image")
                         await MainActor.run {
                             self.currentImage = img
                         }
                         try await room.sendImage(image: img, withBlurhash: false)
+                    }
+                    else {
+                        CirclesApp.logger.debug("Unknown content types: \(item.supportedContentTypes)")
                     }
                 }
                 currentItem = nil